+++ src/app/device_classes/epaper/epaper_drivers.cpp
@@ -10,5 +10,6 @@
 #if HAS_EPAPER
 
 #include "device_classes/epaper/drivers/inkplate5v2_driver.cpp"
+#include "device_classes/epaper/drivers/reterminal_e1003_driver.cpp"
 
 #endif // HAS_EPAPER